refactor(type/built-in): ♻️  Split into chunks
